pyproject.toml
[project]
name = "resume-ai"
version = "0.1.0"
description = "Local-first AI-powered resume builder pipeline"
authors = [{ name = "Your Team" }]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7",
    "jinja2>=3.1",
    "typer>=0.12",
    "openai>=1.20",
    "requests>=2.31",
    "python-docx>=1.1",
    "streamlit>=1.40"
]

[project.optional-dependencies]
pdf = ["weasyprint>=61"]
dev = ["pytest>=8.2", "ruff>=0.5"]

[build-system]
requires = ["setuptools>=64", "wheel>=0.43"]
build-backend = "setuptools.build_meta"
